+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/roomlist/RoomListViewModel.kt
@@ -9,7 +9,6 @@ import de.connect2x.trixnity.messenger.util.getOrNull
 import de.connect2x.trixnity.messenger.viewmodel.ViewModelContext
 import de.connect2x.trixnity.messenger.viewmodel.matrixClients
 import de.connect2x.trixnity.messenger.viewmodel.util.ErrorType
-import de.connect2x.trixnity.messenger.viewmodel.util.Initials
 import de.connect2x.trixnity.messenger.viewmodel.util.RoomName
 import de.connect2x.trixnity.messenger.viewmodel.util.isVerified
 import de.connect2x.trixnity.messenger.viewmodel.verification.SelfVerificationTrigger
@@ -68,6 +67,7 @@ interface RoomListViewModelFactory {
         onRoomSelected: (UserId, RoomId) -> Unit,
         onStartCreateNewRoom: (UserId) -> Unit,
         onUserSettingsSelected: () -> Unit,
+        onUserProfileSelected: () -> Unit,
         onOpenAppInfo: () -> Unit,
         onSendLogs: () -> Unit,
         onOpenAccountsOverview: () -> Unit,
@@ -79,10 +79,11 @@ interface RoomListViewModelFactory {
             onRoomSelected,
             onStartCreateNewRoom,
             onUserSettingsSelected,
+            onUserProfileSelected,
             onOpenAppInfo,
             onSendLogs,
             onOpenAccountsOverview,
-            onAccountSelected
+            onAccountSelected,
         )
     }
 
@@ -140,6 +141,7 @@ class RoomListViewModelImpl(
     private val onRoomSelected: (UserId, RoomId) -> Unit,
     private val onCreateNewRoom: (UserId) -> Unit,
     onUserSettingsSelected: () -> Unit,
+    onUserProfileSelected: () -> Unit,
     onOpenAppInfo: () -> Unit,
     private val onSendLogs: () -> Unit,
     private val onOpenAccountsOverview: () -> Unit,
@@ -172,7 +174,6 @@ class RoomListViewModelImpl(
             .stateIn(coroutineScope, WhileSubscribed(), null)
     private val i18n = get<I18n>()
     private val roomName = get<RoomName>()
-    private val initials = get<Initials>()
 
     override val unverifiedAccounts = viewModelContext.matrixClients
         .flatMapLatest { clients ->
@@ -200,6 +201,7 @@ class RoomListViewModelImpl(
             onAccountSelected = { onAccountSelected() },
             onUserSettingsSelected = onUserSettingsSelected,
             onShowAppInfo = onOpenAppInfo,
+            onShowProfile = onUserProfileSelected,
         )
 
     private val roomListElementViewModels = mutableMapOf<RoomId, RoomListElementViewModel>()