+++ trixnity-messenger/src/commonTest/kotlin/de/connect2x/trixnity/messenger/viewmodel/MainViewModelTest.kt
@@ -729,6 +729,7 @@ class MainViewModelTest : ShouldSpec() {
                                     onRoomSelected: (UserId, RoomId) -> Unit,
                                     onStartCreateNewRoom: (UserId) -> Unit,
                                     onUserSettingsSelected: () -> Unit,
+                                    onUserProfileSelected: () -> Unit,
                                     onOpenAppInfo: () -> Unit,
                                     onSendLogs: () -> Unit,
                                     onOpenAccountsOverview: () -> Unit,
@@ -762,10 +763,13 @@ class MainViewModelTest : ShouldSpec() {
                                             override fun selectActiveAccount(userId: UserId?) {
                                             }
 
-                                            override fun userSettings() {
+                                            override fun openUserSettings() {
                                             }
 
-                                            override fun appInfo() {
+                                            override fun openUserProfile() {
+                                            }
+
+                                            override fun openAppInfo() {
                                             }
                                         }